+++ hoot-core/src/main/cpp/hoot/core/algorithms/linearreference/WayLocation.cpp
@@ -82,7 +82,7 @@ WayLocation::WayLocation(ConstOsmMapPtr map, ConstWayPtr way, double distance)
       _segmentFraction = 0.0;
       for (size_t i = 1; i < way->getNodeCount(); i++)
       {
-        ConstNodePtr n = _map->getNode(_way->getNodeId(i));
+        ConstNodePtr n = _map->getNode(_way->getNodeId(static_cast<int>(i)));
         Coordinate next = n->toCoordinate();
         double delta = next.distance(last);
         last = next;
